// Online_Pharmacy_and_Delivery_Frontend/src/shared/types/api/admin.d.ts
export interface components {
    schemas: {
        PrescriptionReviewDto: {
            decision: string;
            rejectionReason?: string;
        };
        ApiResponsePrescriptionResponseDto: {
            success?: boolean;
            message?: string;
            data?: components["schemas"]["PrescriptionResponseDto"];
            timestamp?: string;
        };
        PrescriptionResponseDto: {
            /** Format: int64 */
            id?: number;
            /** Format: int64 */
            userId?: number;
            userEmail?: string;
            fileUrl?: string;
            fileType?: string;
            status?: string;
            doctorName?: string;
            doctorRegNumber?: string;
            /** Format: int64 */
            reviewedByAdminId?: number;
            rejectionReason?: string;
            /** Format: int64 */
            orderId?: number;
            uploadedAt?: string;
            reviewedAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}